Play-song seperating into components and hooks
