+++ src/pages/api/song/[id]/index.ts
@@ -4,14 +4,22 @@ import { firestore } from "lib/firebase";
 export default async (req: NextApiRequest, res: NextApiResponse) => {
   const { id } = <{ id: string }>req.query;
 
-  try {
-    const document = await firestore.collection("songs").doc(id).get();
-    const data = document.data();
+  const songsCollection = firestore.collection("songs");
+  const commentsCollection = songsCollection.doc(id).collection("comments");
 
-    if (!data)
+  try {
+    const document = await songsCollection.doc(id).get();
+    const song = document.data();
+
+    const { docs } = await commentsCollection.get();
+    const comments = docs.map((doc) => doc.data());
+
+    if (!song)
       return res.status(404).json({ message: "No song found with that id" });
 
-    res.status(200).json(data);
+    song.comments = comments;
+
+    res.status(200).json(song);
   } catch (error) {
     res.status(500).json(error);
   }
