+++ src/types.d.ts
@@ -4,6 +4,7 @@ interface Song {
   artist: string;
   artworkURL: string;
   audioURL: string;
+  comments: Comments;
 }
 
 interface Playlist {
@@ -11,5 +12,11 @@ interface Playlist {
   songs: Songs;
 }
 
+interface Comment {
+  postedBy: string;
+  text: string;
+}
+
 type Songs = Song[];
 type Playlists = Playlist[];
+type Comments = Comment[];
